+++ db.js
@@ -1,6 +1,5 @@
 const { MongoClient } = require('mongodb');
 
-
 const uri = `mongodb+srv://${process.env.MDB_USER}:${process.env.MDB_PASSWORD}@gatherbot.jeulp.mongodb.net/gatherbot?retryWrites=true&w=majority`;
 const client = new MongoClient(uri, { useNewUrlParser: true, useUnifiedTopology: true });
